Making the transpiler more modular
